pyproject.toml
[project]
name = "assistant-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.121.2",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.2.13",
    "pydantic-ai-slim[openai]>=1.25.1",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "ptpython>=3.0.31",
    "pytest>=9.0.1",
]
